py/openai-wss/requirements.txt
websockets
nats-py
